[Update]bigint

+++ db/migrate/20230107061526_create_post_tags.rb
@@ -2,7 +2,7 @@ class CreatePostTags < ActiveRecord::Migration[6.1]
   def change
     create_table :post_tags do |t|
       t.references :post, type: :integer, foreign_key: true
-      t.references :tag,type: :integer, foreign_key: true
+      t.references :tag, type: :bigint, foreign_key: true
       t.timestamps
     end
     # 同じタグを２回保存できないようにする
